Polygon::getNumInteriorRing() return size_t, Polygon::interiorRingN() takes size_t.
git-svn-id: http://svn.osgeo.org/geos/trunk@1630 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/headers/geos/geom/Polygon.h
@@ -105,10 +105,10 @@ public:
 	const LineString* getExteriorRing() const;
 
 	/// Returns number of interior rings (hole)
-	int getNumInteriorRing() const;
+	size_t getNumInteriorRing() const;
 
 	/// Get nth interior ring (hole)
-	const LineString* getInteriorRingN(int n) const;
+	const LineString* getInteriorRingN(size_t n) const;
 
 	std::string getGeometryType() const;
 	virtual GeometryTypeId getGeometryTypeId() const;
@@ -185,6 +185,9 @@ private:
 
 /**********************************************************************
  * $Log$
+ * Revision 1.7  2006/06/08 17:58:57  strk
+ * Polygon::getNumInteriorRing() return size_t, Polygon::interiorRingN() takes size_t.
+ *
  * Revision 1.6  2006/05/04 15:49:39  strk
  * updated all Geometry::getDimension() methods to return Dimension::DimensionType (closes bug#93)
  *
